Expose function to validate migration files
The 'consecutive IDs' check now runs earlier, after file loading.

Fixes #51
